Final Audio and Performance Optimizations
Address all remaining priority recommendations from analysis:
- Hoist Arpeggiator and Euclidean Sequencer to control-rate for massive performance win.
- Fix filter envelope under-sampling (update every 4 samples).
- Implement linear interpolation for modulated Delay lines to eliminate zipper noise.
- Implement two-stage input diffusion (allpass) for FDN Reverb to smooth transients.
- Cache panning coefficients in Voice class to avoid redundant trig calls.
- Improve mixing logic for Resonator and Comb Filter to respect mix parameters.
- Add output limiting (fastTanh) to Resonator and normalization to Granular exciter.
- Replace Moog/Ladder filter fastTanh stages with cheaper polynomial clippers.
- Increase Audio Task priority and fix SDManager resource leak on re-init.
- Ensure thread safety for all critical parameter updates.

+++ Effects.h
@@ -30,7 +30,7 @@ private:
     int16_t bufferL_[MAX_DELAY_SAMPLES];  // 16-bit saves 50% memory
     int16_t bufferR_[MAX_DELAY_SAMPLES];
     uint16_t writePos_;
-    uint16_t delaySamples_;
+    float delaySamples_;
     
     float delayTime_;
     float feedback_;
